get_release and set_version are now in external script release_det.py. IN PROGRESS
